Initial commit of psj project
This is the raw source of psj; I don't have revision history that I can easilly
push. There isn't an installer (yet). It's simply the psj script and all the
modules it depends on. I'll eventually make it into your standard CPAN type
module.

This code works well in my environment. It may not work well in yours. Either
way, have fun with it.